+++ src/club_existing/definitions.d.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/quotes */
+/* eslint-disable @typescript-eslint/semi */
 /* eslint-disable @typescript-eslint/ban-types */
 /* eslint-disable @typescript-eslint/indent */
 
@@ -67,13 +68,6 @@ type ExtendedArchetype = "modular" | "typed" | "vibrating" | "variableheight";
 
 type TypedItemChatSetting = "toOnly" | "fromTo" | "silent";
 type ModularItemChatSetting = "perModule" | "perOption";
-type CommonChatTags =
-	| "SourceCharacter"
-	| "DestinationCharacter"
-	| "DestinationCharacterName"
-	| "TargetCharacter"
-	| "TargetCharacterName"
-	| "AssetName";
 
 type NotificationAudioType = 0 | 1 | 2;
 type NotificationAlertType = 0 | 1 | 3 | 2;
@@ -88,7 +82,11 @@ type VibratorModeSet = "Standard" | "Advanced";
 
 type VibratorModeState = "Default" | "Deny" | "Orgasm" | "Rest";
 
+type VibratorMode = "Off" | "Low" | "Medium" | "High" | "Maximum" | "Random" | "Escalate" | "Tease" | "Deny" | "Edge";
+
 type VibratorRemoteAvailability = "Available" | "NoRemote" | "NoRemoteOwnerRuleActive" | "NoLoversRemote" | "RemotesBlocked" | "CannotInteract" | "NoAccess" | "InvalidItem";
+
+type ItemVulvaFuturisticVibratorAccessMode = "" | "ProhibitSelf" | "LockMember";
 
 /**
  * @property Freeze - Prevents walking and kneeling unaided. There's a few caveats with the kneeling part.
@@ -229,7 +227,7 @@ type AssetGroupItemName =
 	'ItemMouth3' | 'ItemNeck' | 'ItemNeckAccessories' | 'ItemNeckRestraints' |
 	'ItemNipples' | 'ItemNipplesPiercings' | 'ItemNose' | 'ItemPelvis' |
 	'ItemTorso' | 'ItemTorso2' | 'ItemVulva' | 'ItemVulvaPiercings' |
-	'ItemHandheld' |
+	'ItemHandheld' | 'ItemScript' |
 
 	'ItemHidden' /* TODO: investigate, not a real group */
 	;
@@ -271,8 +269,15 @@ type CraftingPropertyType =
 	"Flexible" | "Nimble" | "Arousing" | "Dull"
 	;
 
-/** Enum values for {@link CraftingStatusCode}. */
+type AssetAttribute =
+	"Skirt" |
+	"ShortHair" |
+	"CanAttachMittens"
+	;
+
 type CraftingStatusType = 0 | 1 | 2;
+
+type ItemColorMode = "Default" | "ColorPicker";
 
 //#endregion
 
@@ -388,17 +393,235 @@ type MessageActionType = "Action" | "Chat" | "Whisper" | "Emote" | "Activity" | 
 
 type MessageContentType = string;
 
-interface ChatMessageDictionaryEntry {
-	Tag?: CommonChatTags | string;
-	Text?: string;
-	MemberNumber?: number;
-	TextToLookUp?: string;
-	AssetName?: string;
-	AssetGroupName?: string;
-	Automatic?: boolean;
-	ShockIntensity?: number;
-	ActivityCounter?: number;
+type CharacterReferenceTag =
+	| "SourceCharacter"
+	| "DestinationCharacter"
+	| "DestinationCharacterName"
+	| "TargetCharacter"
+	| "TargetCharacterName"
+
+type CommonChatTags =
+	| CharacterReferenceTag
+	| "AssetName";
+
+/**
+ * A dictionary entry containing a replacement tag to be replaced by some value. The replacement strategy depends on
+ * the type of dictionary entry.
+ */
+interface TaggedDictionaryEntry {
+	/** The tag that will be replaced in the message */
+	Tag: string;
 }
+
+/**
+ * A dictionary entry used to reference a character. The character reference tag will be replaced with the provided
+ * character's name or pronoun. The display format will depend on the tag chosen.
+ * Example substitutions for each tag (assuming the character name is Ben987):
+ * * SourceCharacter: "Ben987"
+ * * DestinationCharacter: "Ben987's" (if character is not self), "her"/"him" (if character is self)
+ * * DestinationCharacterName: "Ben987's"
+ * * TargetCharacter: "Ben987" (if character is not self), "herself"/"himself" (if character is self)
+ * * TargetCharacterName: "Ben987"
+ * @deprecated Use {@link SourceCharacterDictionaryEntry} and {@link TargetCharacterDictionaryEntry} instead.
+ */
+interface CharacterReferenceDictionaryEntry extends TaggedDictionaryEntry {
+	/** The member number of the referenced character */
+	MemberNumber: number;
+	/** The character reference tag, determining how the character's name or pronoun will be interpreted */
+	Tag: CharacterReferenceTag;
+	/**
+	 * The nickname of the referenced character
+	 * @deprecated Redundant information
+	 */
+	Text?: string;
+}
+
+/**
+ * A dictionary entry used to indicate the source character of a chat message or action (i.e. the character initiating
+ * the message or action).
+ */
+interface SourceCharacterDictionaryEntry {
+	SourceCharacter: number;
+}
+
+/**
+ * A dictionary entry used to indicate the target character of a chat message or action (i.e. the character that is
+ * being acted upon as part of the message or action).
+ */
+interface TargetCharacterDictionaryEntry {
+	TargetCharacter: number;
+}
+
+/**
+ * A dictionary entry which indicates the focused group. This represents the group that was focused or interacted with
+ * when sending a chat message. For example, if the message was caused by performing an activity or modifying an item
+ * on the `ItemArms` group, then it would be appropriate to send this dictionary entry with `ItemArms` as the focus
+ * group name.
+ */
+interface FocusGroupDictionaryEntry {
+	/**
+	 * The tag to be replaced - this is always FocusAssetGroup.
+	 * @deprecated Redundant information.
+	 */
+	Tag?: "FocusAssetGroup";
+	/** The group name representing focused group for the purposes of the sent message */
+	FocusGroupName: AssetGroupName;
+}
+
+/**
+ * A direct text substitution dictionary entry. Any occurrences of the given {@link Tag} string in the associated
+ * message will be directly replaced with the {@link Text} from this dictionary entry (no text lookup will be done).
+ * For example, given the message:
+ * ```
+ * Life is like a box of ConfectionaryName.
+ * ```
+ * and the {@link TextDictionaryEntry}:
+ * ```js
+ * {Tag: "ConfectionaryName", Text: "chocolates"}
+ * ```
+ * The resulting message would be:
+ * ```
+ * Life is like a box of chocolates.
+ * ```
+ */
+interface TextDictionaryEntry extends TaggedDictionaryEntry {
+	/** The text that will be substituted for the tag */
+	Text: string;
+}
+
+/**
+ * A text substitution dictionary entry with text lookup functionality. Any occurrences of the given {@link Tag} string
+ * in the associated message will be replaced with the {@link Text} from the dictionary entry, but only after a text
+ * lookup has been done on the {@link Text}, meaning that if the text has localisations, the localised version will be
+ * used. The text will be looked up against `Dialog_Player.csv`.
+ * For example, given the message:
+ * ```
+ * Hello, {GreetingObjectName}!
+ * ```
+ * And the {@link TextLookupDictionaryEntry}:
+ * ```js
+ * {Tag: "GreetingObjectName", TextToLookup: "WorldObject"}
+ * ```
+ * And the following in `Dialog_Player.csv`:
+ * ```
+ * WorldObject,,,World,,
+ * ```
+ * The text to lookup (`"WorldObject"`) would be looked up against `Dialog_Player.csv`, resolving to `"World"`. This
+ * would then be used to replace the tag `"GreetingObjectName"` in the message, resulting in:
+ * ```
+ * Hello, World!
+ * ```
+ */
+interface TextLookupDictionaryEntry extends TaggedDictionaryEntry {
+	/** The text whose lookup will be substituted for the tag */
+	TextToLookUp: string;
+}
+
+/**
+ * A dictionary entry that references an asset group. Note that this is different from
+ * {@link FocusGroupDictionaryEntry}, which denotes the group being acted on. A dictionary should only ever contain
+ * one {@link FocusGroupDictionaryEntry}, whereas it may contain many {@link GroupReferenceDictionaryEntry}s. This
+ * represents any group that might be referenced in the message, but is not necessarily the focused group.
+ * For example, given the message:
+ * ```
+ * Use your BodyPart!
+ * ```
+ * And the {@link GroupReferenceDictionaryEntry}:
+ * ```
+ * {Tag: "BodyPart", GroupName: "ItemHands"}
+ * ```
+ * The name of the `"ItemHands"` group would be looked up, and this would be used to replace the `"BodyPart"` tag. The
+ * resulting message would be:
+ * ```
+ * Use your Hands!
+ * ```
+ */
+interface GroupReferenceDictionaryEntry extends TaggedDictionaryEntry {
+	/** The name of the asset group to reference */
+	GroupName: AssetGroupName;
+}
+
+/**
+ * A dictionary entry that references an asset. Note that a dictionary may contain multiple of these entries, one for
+ * each asset mentioned or referenced in the message. For example, a message when swapping two restraints might contain
+ * two of these entries, one for the restraint being removed, and one for the restraint being added.
+ */
+interface AssetReferenceDictionaryEntry extends GroupReferenceDictionaryEntry {
+	/** The name of the asset being referenced */
+	AssetName: string;
+}
+
+/**
+ * A special instance of an {@link AssetReferenceDictionaryEntry} which indicates that this asset was used to carry
+ * out an activity.
+ */
+interface ActivityAssetReferenceDictionaryEntry extends AssetReferenceDictionaryEntry {
+	Tag: "ActivityAsset";
+}
+
+/**
+ * A metadata dictionary entry sent with a shock event message including a shock intensity representing the strength
+ * of the shock. This is used to determine the severity of any visual or gameplay effects the shock may have.
+ */
+interface ShockEventDictionaryEntry {
+	/** The intensity of the shock - must be a non-negative number */
+	ShockIntensity: number;
+}
+
+/**
+ * A metadata dictionary entry indicating that the message has been generated due to an automated event. Can be used
+ * to filter out what might otherwise be spammy chat messages (these include things like automatic vibrator intensity
+ * changes and events & messages triggered by some futuristic items).
+ */
+interface AutomaticEventDictionaryEntry {
+	/** Indicates that this message was triggered by an automatic event */
+	Automatic: true;
+}
+
+/**
+ * A metadata dictionary entry carrying a numeric counter for an associated event or activity. Currently only used by
+ * the Anal Beads XL to indicate how many beads were inserted.
+ */
+interface ActivityCounterDictionaryEntry {
+	/** Counter metadata to be sent with a message */
+	ActivityCounter: number;
+}
+
+/**
+ * A dictionary entry for group lookup & replacement. Used ambiguously for both {@link FocusGroupDictionaryEntry} and
+ * {@link GroupReferenceDictionaryEntry}. This dictionary entry type is deprecated, and one of the aforementioned entry
+ * types should be used instead.
+ * @deprecated Use {@link FocusGroupDictionaryEntry}/{@link GroupReferenceDictionaryEntry}
+ */
+interface AssetGroupNameDictionaryEntry {
+	Tag?: "FocusAssetGroup";
+	AssetGroupName: AssetGroupName;
+}
+
+/**
+ * A dictionary entry indicating the name of an activity. Sent with chat messages to indicate that an activity was
+ * carried out as part of the message.
+ */
+interface ActivityNameDictionaryEntry {
+	/** The name of the activity carried out */
+	ActivityName: string;
+}
+
+type ChatMessageDictionaryEntry =
+	| CharacterReferenceDictionaryEntry
+	| SourceCharacterDictionaryEntry
+	| TargetCharacterDictionaryEntry
+	| FocusGroupDictionaryEntry
+	| TextDictionaryEntry
+	| TextLookupDictionaryEntry
+	| GroupReferenceDictionaryEntry
+	| AssetReferenceDictionaryEntry
+	| ActivityAssetReferenceDictionaryEntry
+	| ShockEventDictionaryEntry
+	| AutomaticEventDictionaryEntry
+	| ActivityCounterDictionaryEntry
+	| AssetGroupNameDictionaryEntry
+	| ActivityNameDictionaryEntry;
 
 type ChatMessageDictionary = ChatMessageDictionaryEntry[];
 
@@ -431,19 +654,23 @@ interface IChatRoomMessageMetadata {
 	TargetMemberNumber?: number;
 	/** Whether the message is considered game-initiated. Used for automatic vibe changes for example. */
 	Automatic?: boolean;
-	/** The group the message applies to */
-	GroupName?: string;
+	/** The group that has been interacted with to trigger the message */
+	FocusGroup?: AssetGroup;
+	/** The name of the group that has been interacted with to trigger the message */
+	GroupName?: AssetGroupName;
+	/** The assets referenced in the message */
+	Assets?: Record<string, Asset>;
+	/** The groups referenced in the message */
+	Groups?: Record<string, AssetGroup>;
 	/** How intense the shock should be */
 	ShockIntensity?: number;
 	ActivityCounter?: number;
 	/** The triggered activity */
 	ActivityName?: string;
-	/** The group where the activity is triggered */
-	ActivityGroup?: string;
 	/** The name of the asset used for the activity */
-	ActivityAsset?: string;
-	/** The group the asset used is in */
-	ActivityAssetGroup?: string;
+	ActivityAsset?: Asset;
+	/** The name of the chatroom, appropriately garbled */
+	ChatRoomName?: string;
 }
 
 /**
@@ -553,7 +780,7 @@ interface AssetGroup {
 	Description: string;
 	Asset: Asset[];
 	ParentGroupName: string;
-	Category: 'Appearance' | 'Item';
+	Category: 'Appearance' | 'Item' | 'Script';
 	IsDefault: boolean;
 	IsRestraint: boolean;
 	AllowNone: boolean;
@@ -646,6 +873,10 @@ interface AssetLayer {
 	GroupAlpha?: AlphaDefinition[];
 	/** A module for which the layer can have types. */
 	ModuleType: string[] | null;
+	/* Specifies that this layer should not be drawn if the character is wearing any item with the given attributes */
+	HideForAttribute: AssetAttribute[] | null;
+	/* Specifies that this layer should not be drawn unless the character is wearing an item with one of the given attributes */
+	ShowForAttribute: AssetAttribute[] | null;
 }
 
 /** An object defining a group of alpha masks to be applied when drawing an asset layer */
@@ -708,7 +939,7 @@ interface Asset {
 	Hide?: string[];
 	HideItem?: string[];
 	HideItemExclude: string[];
-	HideItemAttribute: string[];
+	HideItemAttribute: AssetAttribute[];
 	Require: string[];
 	SetPose?: string[];
 	AllowPose: string[] | null;
@@ -747,6 +978,8 @@ interface Asset {
 	RemoveItemOnRemove: { Name: string; Group: string; Type?: string; }[];
 	AllowEffect?: EffectName[];
 	AllowBlock?: AssetGroupItemName[];
+	AllowHide?: AssetGroupItemName[];
+	AllowHideItem?: string[];
 	AllowType?: string[];
 	DefaultColor?: ItemColor;
 	Opacity: number;
@@ -764,7 +997,6 @@ interface Asset {
 	DynamicDescription: (C: Character) => string;
 	DynamicPreviewImage: (C: Character) => string;
 	DynamicAllowInventoryAdd: (C: Character) => boolean;
-	DynamicExpressionTrigger: (C: Character) => ExpressionTrigger[] | null | undefined;
 	DynamicName: (C: Character) => string;
 	DynamicGroupName: AssetGroupName;
 	DynamicActivity: (C: Character) => string | null | undefined;
@@ -788,7 +1020,7 @@ interface Asset {
 	Layer: AssetLayer[];
 	ColorableLayerCount: number;
 	Archetype?: string;
-	Attribute: string[];
+	Attribute: AssetAttribute[];
 	PreviewIcons: InventoryIcon[];
 	Tint: TintDefinition[];
 	AllowTint: boolean;
@@ -797,6 +1029,8 @@ interface Asset {
 	CraftGroup: string;
 	ColorSuffix: Record<string, string>;
 	ExpressionPrerequisite?: string[];
+	TextMaxLength: null | Partial<Record<PropertyTextNames, number>>;
+	TextFont: null | string;
 }
 
 //#endregion
@@ -834,7 +1068,7 @@ interface Activity {
 	TargetSelf?: AssetGroupItemName[] | true;
 	/** Whether to reverse the prerequisite checks for that one */
 	Reverse?: true;
-	/** used for setting AutoPunishGagActionFlag */
+	/** used for setting {@link ExtendedItemAutoPunishHandled} */
 	MakeSound?: boolean;
 	/** An action that trigger when that activity is used */
 	StimulationAction?: StimulationAction;
@@ -960,6 +1194,29 @@ interface ScreenFunctions {
 
 //#region Characters
 
+/** A struct for representing an item with special permissions (limited, favorited, etc). */
+interface ItemPermissions {
+	/** The {@link Asset.Name} of the item */
+	Name: string;
+	/** The {@link AssetGroup.Name} of the item */
+	Group: AssetGroupName;
+	/**
+	 * Either the item's {@link ItemProperties.Type} or, in the case of modular items,
+	 * a substring thereof denoting the type of a single module
+	 */
+	Type?: string | null;
+}
+
+interface ScriptPermission {
+	permission: number;
+}
+
+type ScriptPermissionProperty = "Hide" | "Block";
+
+type ScriptPermissionLevel = "Self" | "Owner" | "Lovers" | "Friends" | "Whitelist" | "Public";
+
+type ScriptPermissions = Record<ScriptPermissionProperty, ScriptPermission>;
+
 interface Character {
 	ID: number;
 	/** Only on `Player` */
@@ -982,15 +1239,16 @@ interface Character {
 	Pose: string[];
 	Effect: string[];
 	Tints: ResolvedTintDefinition[];
+	Attribute: AssetAttribute[];
 	FocusGroup: AssetGroup | null;
 	Canvas: HTMLCanvasElement | null;
 	CanvasBlink: HTMLCanvasElement | null;
 	MustDraw: boolean;
 	BlinkFactor: number;
 	AllowItem: boolean;
-	BlockItems: any[];
-	FavoriteItems: any[];
-	LimitedItems: any[];
+	BlockItems: ItemPermissions[];
+	FavoriteItems: ItemPermissions[];
+	LimitedItems: ItemPermissions[];
 	WhiteList: number[];
 	HeightModifier: number;
 	MemberNumber?: number;
@@ -1037,7 +1295,7 @@ interface Character {
 	IsLoverOfPlayer: () => boolean;
 	GetLoversNumbers: (MembersOnly?: boolean) => (number | string)[];
 	AllowedActivePose: string[];
-	HiddenItems: any[];
+	HiddenItems: ItemPermissions[];
 	HeightRatio: number;
 	HasHiddenItems: boolean;
 	SavedColors: HSVColor[];
@@ -1060,6 +1318,7 @@ interface Character {
 	IsLoverPrivate: () => boolean;
 	IsEdged: () => boolean;
 	IsPlayer: () => this is PlayerCharacter;
+	IsBirthday: () => boolean;
 	IsOnline: () => boolean;
 	IsNpc: () => boolean;
 	IsSimple: () => boolean;
@@ -1071,6 +1330,7 @@ interface Character {
 	HasEffect: (Effect: string) => boolean;
 	HasTints: () => boolean;
 	GetTints: () => RGBAColor[];
+	HasAttribute: (attribute: AssetAttribute) => boolean;
 	DrawPose?: string[];
 	DrawAppearance?: Item[];
 	AppearanceLayers?: AssetLayer[];
@@ -1118,6 +1378,8 @@ interface Character {
 		DisablePickingLocksOnSelf: boolean;
 		GameVersion: string;
 		ItemsAffectExpressions: boolean;
+		ScriptPermissions: ScriptPermissions;
+		WheelFortune: string;
 	};
 	Game?: {
 		LARP?: GameLARPParameters,
@@ -1268,6 +1530,8 @@ interface PlayerCharacter extends Character {
 		DisableAnimations?: any;
 		/** @deprecated */
 		SearchShowsFullRooms?: any;
+		CensoredWordsList: string;
+		CensoredWordsLevel: number;
 	};
 	VisualSettings?: {
 		ForceFullHeight?: boolean;
@@ -1429,10 +1693,154 @@ interface NPCTrait {
 
 //#region Extended items
 
+/** A struct-type that maps archetypes to their respective extended item data.  */
+interface ExtendedDataLookupStruct {
+	[ExtendedArchetype.TYPED]: TypedItemData;
+	[ExtendedArchetype.MODULAR]: ModularItemData;
+	[ExtendedArchetype.VIBRATING]: VibratingItemData;
+	[ExtendedArchetype.VARIABLEHEIGHT]: VariableHeightData;
+}
+
 interface AssetOverrideHeight {
 	Height: number;
 	Priority: number;
 	HeightRatioProportion?: number;
+}
+
+/**
+ * Base properties of extended items derived from their respective {@link Asset} definition.
+ *
+ * Those are the properties the main game code enforces.
+ */
+interface AssetDefinitionProperties {
+	/**
+	 * The difficulty of the item
+	 * @see {@link Asset.Difficulty}
+	 */
+	Difficulty?: number;
+	/**
+	 * ???
+	 * @see {@link Asset.Attribute}
+	 */
+	Attribute?: AssetAttribute[];
+
+	/**
+	 * Override the height of the item
+	 * @see {@link Asset.OverrideHeight}
+	 */
+	OverrideHeight?: AssetOverrideHeight;
+	/**
+	 * How much the character should be moved up
+	 * @see {@link Asset.HeightModifier}
+	 */
+	HeightModifier?: number;
+	/**
+	 * The drawing priority of the item
+	 * @see {@link Asset.OverridePriority}
+	 */
+	OverridePriority?: number;
+	/**
+	 * The default color of the item
+	 * @see {@link Asset.DefaultColor}
+	 */
+	DefaultColor?: ItemColor;
+
+	/**
+	 * A list of allowed activities
+	 * @see {@link Asset.AllowActivity}
+	 */
+	AllowActivity?: string[];
+	/**
+	 * A list of groups allowed activities
+	 * @see {@link Asset.AllowActivityOn}
+	 */
+	AllowActivityOn?: AssetGroupName[];
+
+	/**
+	 * Items that should be hidden by this item
+	 * @see {@link Asset.HideItem}
+	 */
+	HideItem?: string[];
+	/**
+	 * Items that should not be hidden by this item
+	 * @see {@link Asset.HideItemExclude}
+	 */
+	HideItemExclude?: string[];
+	/**
+	 * Items groups that should be hidden by this item
+	 * @see {@link Asset.Hide}
+	 */
+	Hide?: AssetGroupName[];
+
+	/**
+	 * The groups that this item blocks
+	 * @see {@link Asset.Block}
+	 */
+	Block?: AssetGroupItemName[];
+
+	/**
+	 * Effects that are applied by this item
+	 * @see {@link Asset.Effect}
+	 */
+	Effect?: EffectName[];
+
+	/**
+	 * A list of custom tints
+	 * @see {@link Asset.Tint}
+	 */
+	Tint?: TintDefinition[];
+
+	// Pose-related properties
+
+	/**
+	 * A list of poses that should forcefully be set
+	 * @see {@link Asset.SetPose}
+	 */
+	SetPose?: AssetPoseName[];
+	/**
+	 * A list of poses
+	 * @see {@link Asset.AllowActivePose}
+	 */
+	AllowActivePose?: AssetPoseName[];
+	/**
+	 * A list of allowed poses
+	 * @see {@link Asset.AllowPose}
+	 */
+	AllowPose?: AssetPoseName[];
+	/**
+	 * A list of poses
+	 * @see {@link Asset.WhitelistActivePose}
+	 */
+	WhitelistActivePose?: AssetPoseName[];
+	/**
+	 * A list of poses that should be frozen
+	 * @see {@link Asset.FreezeActivePose}
+	 */
+	FreezeActivePose?: AssetPoseName[];
+
+	/**
+	 * Whether an item can be unlocked by the player even if they're restrained
+	 * @see {@link Asset.SelfUnlock}
+	 */
+	SelfUnlock?: boolean;
+
+	/**
+	 * The timer for after how long until a lock should be removed
+	 * @see {@link Asset.RemoveTimer}
+	 */
+	RemoveTimer?: number;
+
+	/**
+	 * The asset's draw opacity
+	 * @see {@link Asset.Opacity}
+	 */
+	Opacity?: number;
+
+	/**
+	 * A custom background for this option that overrides the default
+	 * @see {@link Asset.CustomBlindBackground}
+	 */
+	CustomBlindBackground?: string;
 }
 
 /**
@@ -1441,43 +1849,22 @@ interface AssetOverrideHeight {
  * Those are the properties the main game code enforces.
  */
 interface ItemPropertiesBase {
+	/** A string (or `null`) denoting the state of an extended item. How the type-string translate to concrete properties depends on the Archetype in question. */
 	Type?: string | null;
+
+	/** A facial expression */
 	Expression?: string;
-	Difficulty?: number;
-	OverrideHeight?: AssetOverrideHeight;
-	HeightModifier?: number;
-	OverridePriority?: unknown;
-	DefaultColor?: string;
 
-	Attribute?: string[];
-
-	AllowActivity?: string[];
-	AllowActivityOn?: AssetGroupName[];
-
-	/** Items hidden by this one */
-	HideItem?: string[];
-	HideItemExclude?: string[];
-	Hide?: AssetGroupName[];
-
-	/** The groups that this item blocks */
-	Block?: AssetGroupItemName[];
-
-	Effect?: EffectName[];
+	/** Whether the asset affects should be overriden rather than extended */
 	OverrideAssetEffect?: boolean;
 
-	Tint?: TintDefinition[];
+	// Vibratory-related properties
 
-	/* Pose-related properties */
-
-	SetPose?: AssetPoseName[];
-	AllowActivePose?: AssetPoseName[];
-	AllowPose?: AssetPoseName[];
-	WhitelistActivePose?: AssetPoseName[];
-	FreezeActivePose?: AssetPoseName[];
-
-	/* Vibratory-related properties */
+	/** The vibrator mode */
 	Mode?: VibratorMode;
+	/** The vibrator intensity */
 	Intensity?: VibratorIntensity;
+	/** The vibrator's state; only relevant for advanced vibrator modes */
 	State?: VibratorModeState;
 }
 
@@ -1488,18 +1875,18 @@ interface ItemPropertiesBase {
  * per-item.
  */
 interface ItemPropertiesCustom {
+	/** The member number of the player adding the item */
 	ItemMemberNumber?: number;
 
+	/** The member number of the player adding the item */
 	MemberNumber?: number;
 
-	SelfUnlock?: boolean;
-
 	//#region Lock properties
+
 	/** Asset name of the lock */
 	LockedBy?: AssetLockType;
+	/** The member number of the person that applied the lock */
 	LockMemberNumber?: number | string;
-	/** @see BC_Asset.MaxTimer */
-	RemoveTimer?: number;
 	/** `/^[A-Z]{1,8}$/`, Used by `PasswordPadlock`, `SafewordPadlock` and `TimerPasswordPadlock` lock */
 	Password?: string;
 	/** Comma separated numbers */
@@ -1522,80 +1909,107 @@ interface ItemPropertiesCustom {
 	EnableRandomInput?: boolean;
 	/** List of people who publicly modified time on lock; used by `LoversTimerPadlock`, `MistressTimerPadlock`, `OwnerTimerPadlock`, `TimerPasswordPadlock` */
 	MemberNumberList?: number[];
+
 	//#endregion
 
-	InflateLevel?: number;
+	/** The inflation level of inflatable items */
+	InflateLevel?: 0 | 1 | 2 | 3 | 4;
 
-	SuctionLevel?: number;
+	/** The suction level of items with a suction effect */
+	SuctionLevel?: 0 | 1 | 2 | 3 | 4;
 
 	/** 1st line of text for user-entered text data */
 	Text?: string;
 	/** 2nd line of text for user-entered text data */
 	Text2?: string;
+	/** 3rd line of text for user-entered text data */
+	Text3?: string;
 
+	/** Whether the item blocks access to the butt */
 	LockButt?: boolean;
 
-	/* Futuristic Set open permissions */
+	// #region Futuristic Set open permissions
 
+	/** Whether all players can use futuristic head devices */
 	OpenPermission?: boolean;
+	/** Whether all players can use futuristic arm devices */
 	OpenPermissionArm?: boolean;
+	/** Whether all players can use futuristic leg devices */
 	OpenPermissionLeg?: boolean;
+	/** Whether all players can use futuristic chastity devices */
 	OpenPermissionChastity?: boolean;
+	/** Whether the usage of remotes is blocked */
+	BlockRemotes?: boolean;
+
+	// #endregion
 
 	/** The futuristic bra's heart rate value */
 	HeartRate?: number;
 	/** Is the futuristic bra's heart icon shown */
 	HeartIcon?: boolean;
 
-	/* Futuristic gag & panel gag settings */
-	AutoPunish?: number;
+	// #region Futuristic gag & panel gag settings */
+
+	/** The item's auto-punishment sensitivity */
+	AutoPunish?: 0 | 1 | 2 | 3;
+	/** The remaining time for the gag's auto-inflation */
 	AutoPunishUndoTime?: number;
-	AutoPunishUndoTimeSetting?: number;
-	OriginalSetting?: "Padded" | "LightBall" | "Ball" | "Plug";
-	ChatMessage?: boolean;
-	BlinkState?: number;
+	/** The default time for the gag's auto-inflation */
+	AutoPunishUndoTimeSetting?: 120000 | 300000 | 900000 | 3600000 | 72000000;
+	/** The gag module-index prior to triggering auto-inflation */
+	OriginalSetting?: 0 | 1 | 2 | 3;
+	/** Whether gag's blinking light is on or off */
+	BlinkState?: boolean;
+	/**
+	 * An extended item option
+	 * @todo Investigate whether this property still actually exists
+	 */
 	Option?: ExtendedItemOption;
 
-	BlockRemotes?: boolean;
+	// #endregion
 
-	/* Futuristic chastity settings */
+	// #region Futuristic chastity settings
 
+	/** Whether attempting to remove the belt should result in punishment */
 	PunishStruggle?: boolean;
+	/** Whether attempting to remove an item in general should result in punishment */
 	PunishStruggleOther?: boolean;
+	/** Whether orgasms should result in punishment */
 	PunishOrgasm?: boolean;
+	/** Whether standing up should result in punishment */
 	PunishStandup?: boolean;
-	PunishSpeech?: number;
-	PunishRequiredSpeech?: number;
+	/** The punishment for talking; represents an index of {@link FuturisticTrainingBeltSpeechPunishments} */
+	PunishSpeech?: 0 | 1 | 2 | 3;
+	/** The punishment for not speaking a required word; represents an index of {@link FuturisticTrainingBeltSpeechPunishments} */
+	PunishRequiredSpeech?: 0 | 1 | 2 | 3;
+	/** A string with comma-separated required words */
 	PunishRequiredSpeechWord?: string;
-	PunishProhibitedSpeech?: number;
+	/** The punishment for speaking a prohibited word; represents an index of {@link FuturisticTrainingBeltSpeechPunishments} */
+	PunishProhibitedSpeech?: 0 | 1 | 2 | 3;
+	/** A string with comma-separated prohibited words */
 	PunishProhibitedSpeechWords?: string;
+	/** Internal cooldown timer for automatic shocks */
 	NextShockTime?: number;
+	/** The mode of the belts vibrator; represents an index of {@link FuturisticTrainingBeltModes} */
+	PublicModeCurrent?: 0 | 1 | 2 | 3 | 4 | 5 | 6;
+	/** An integer denoting who can access the belt; represents an index of {@link FuturisticTrainingBeltPermissions} */
+	PublicModePermission?: 0 | 1 | 2;
 
-	PublicModeCurrent?: number;
-	PublicModePermission?: number;
+	// #endregion
 
-	/** The futuristic vibrator's trigger words */
+	/** A comma-seperated string with the futuristic vibrator's trigger words */
 	TriggerValues?: string;
-	AccessMode?: string;
+	/** A string denoting who has permission to use the vibrator's trigger words */
+	AccessMode?: ItemVulvaFuturisticVibratorAccessMode;
 
-	/* Pleasure panties settings */
+	/** How intense the shock should be */
 	ShockLevel?: 0 | 1 | 2;
-	LockCrotch?: boolean;
-	OrgasmLock?: number;
 
 	/** The number of inserted beads */
 	InsertedBeads?: 1 | 2 | 3 | 4 | 5;
 
-	/** Whether the item displays a chat message or not */
+	/** Whether the item displays a chat message to all other people in the room */
 	ShowText?: boolean;
-
-	/** How sensitive the item is to whatever its reacting to */
-	Sensitivity?: number;
-
-	/** The asset's draw opacity */
-	Opacity?: number;
-
-	CustomBlindBackground?: string;
 
 	/** Number of times the item was triggered; often used by shock collars */
 	TriggerCount?: number;
@@ -1606,12 +2020,23 @@ interface ItemPropertiesCustom {
 	/** Allows reverting back to these properties on exiting an extended menu */
 	Revert?: boolean;
 
-	/** Door and padding options for the kennel */
+	/** Whether the kennel door is open */
 	Door?: boolean;
+	/** Whether the kennel has padding */
 	Padding?: boolean;
+
+	/** Only available as overrides on the script item */
+	UnHide?: AssetGroupName[];
+
+	/** Lucky Wheel: the section labels */
+	Texts?: string[];
+
+	/** Lucky Wheel: the angle the wheel should spin to */
+	TargetAngle?: number;
+
 }
 
-interface ItemProperties extends ItemPropertiesBase, ItemPropertiesCustom { }
+interface ItemProperties extends ItemPropertiesBase, AssetDefinitionProperties, ItemPropertiesCustom { }
 
 /**
  * An object containing the extended item definition for an asset.
@@ -1681,7 +2106,13 @@ interface ExtendedItemOption {
 	/** If the option has a subscreen, this can set a particular archetype to use */
 	Archetype?: ExtendedArchetype;
 	/** If the option has an archetype, sets the config to use */
-	ArchetypeConfig?: TypedItemConfig | ModularItemConfig | VibratingItemConfig | VariableHeightConfig
+	ArchetypeConfig?: TypedItemConfig | ModularItemConfig | VibratingItemConfig | VariableHeightConfig;
+	/**
+	 * A unique (automatically assigned) identifier of the struct type
+	 * @todo consider making an {@link ExtendedItemOption} struct type wherein this field is mandatory once
+	 * more extended items have been assigned an arhcetype
+	 */
+	OptionType?: "ExtendedItemOption";
 }
 
 /**
@@ -1709,6 +2140,19 @@ interface ExtendedItemChatData<OptionType> {
  */
 type ExtendedItemChatCallback<OptionType> = (
 	chatData: ExtendedItemChatData<OptionType>,
+) => string;
+
+/**
+ * @param {Character} C - The selected NPC
+ * @param {OptionType} Option - The currently selected extended item option
+ * @param {OptionType} PreviousOption - The previously selected extended item option
+ * @returns {string} - The chat prefix that should be used for this type change
+ * @template OptionType
+ */
+type ExtendedItemNPCCallback<OptionType> = (
+	C: Character,
+	Option: OptionType,
+	PreviousOption: OptionType,
 ) => string;
 
 /**
@@ -1766,7 +2210,7 @@ type ModularItemAssetConfig = ExtendedItemAssetConfig<"modular", ModularItemConf
 /** An object defining all of the required configuration for registering a modular item */
 interface ModularItemConfig {
 	/** The module definitions for the item */
-	Modules?: ModularItemModule[];
+	Modules?: ModularItemModuleBase[];
 	/**
 	 * The item's chatroom message setting. Determines the level of
 	 * granularity for chatroom messages when the item's module values change.
@@ -1797,6 +2241,13 @@ interface ModularItemConfig {
 		Exit?: () => void;
 		Validate?: ExtendedItemValidateScriptHookCallback<ModularItemOption>;
 	};
+	/**
+	 * To-be initialized properties independant of the selected item module(s).
+	 * Relevant if there are properties that are (near) exclusively managed by {@link ModularItemConfig.ScriptHooks} functions.
+	 */
+	BaselineProperty?: ItemProperties;
+	/** A boolean indicating whether or not images should be drawn for the module selection screen. */
+	DrawImages?: boolean;
 }
 
 interface ModularItemDialogConfig {
@@ -1822,8 +2273,8 @@ interface ModularItemDialogConfig {
 	ChatPrefix?: string | ExtendedItemChatCallback<ModularItemOption>;
 }
 
-/** An object describing a single module for a modular item. */
-interface ModularItemModule {
+/** A (partially parsed) object describing a single module for a modular item. */
+interface ModularItemModuleBase {
 	/** The name of this module - this is usually a human-readable string describing what the
 	 * module represents (e.g. Straps). It is used for display text keys, and should be unique across all of the modules
 	 * for the item.
@@ -1835,13 +2286,27 @@ interface ModularItemModule {
 	 */
 	Key: string;
 	/** The list of option definitions that can be chosen within this module. */
-	Options: ModularItemOption[];
+	Options: ModularItemOptionBase[];
 	/** Whether or not this module can be selected by the wearer */
 	AllowSelfSelect?: boolean;
+	/** A unique (automatically assigned) identifier of the struct type */
+	OptionType?: "ModularItemModule";
+	/** A boolean indicating whether or not images should be drawn within this particular module. */
+	DrawImages?: boolean;
 }
 
-/** An object describing a single option within a module for a modular item. */
-interface ModularItemOption {
+/** An object describing a single module for a modular item. */
+interface ModularItemModule extends ModularItemModuleBase {
+	/** A unique (automatically assigned) identifier of the struct type */
+	OptionType: "ModularItemModule";
+	/** The list of option definitions that can be chosen within this module. */
+	Options: ModularItemOption[];
+	/** A boolean indicating whether or not images should be drawn within this particular module. */
+	DrawImages: boolean;
+}
+
+/** A (partially parsed) object describing a single option within a module for a modular item. */
+interface ModularItemOptionBase {
 	/** The additional difficulty associated with this option - defaults to 0 */
 	Difficulty?: number;
 	/** The required bondage skill level for this option */
@@ -1884,6 +2349,34 @@ interface ModularItemOption {
 	OverridePriority?: number;
 	/** A list of activities enabled by that module */
 	AllowActivity?: string[];
+	/** The name of the option; automatically set to {@link ModularItemModule.Key} + the option's index */
+	Name?: string;
+	/** A unique (automatically assigned) identifier of the struct type */
+	OptionType?: "ModularItemOption";
+	/** Trigger this expression when changing to this option */
+	Expression?: ExpressionTrigger[];
+}
+
+/** An object describing a single option within a module for a modular item. */
+interface ModularItemOption extends ModularItemOptionBase {
+	/** The name of the option; automatically set to {@link ModularItemModule.Key} + the option's index */
+	Name: string;
+	/** A unique (automatically assigned) identifier of the struct type */
+	OptionType: "ModularItemOption";
+}
+
+/** A struct with drawing data for a given module. */
+interface ModularItemDrawData {
+	/** The number of pages */
+	pageCount: number,
+	/** Whether pagination is required; i.e. if the number of buttons is larger than {@link ModularItemDrawData.itemsPerPage} */
+	paginate: boolean,
+	/** An array with two-tuples of X and Y coordinates for the buttons */
+	positions: [number, number][],
+	/** Whether each button should be accompanied by a preview image */
+	drawImages: boolean,
+	/** The number of buttons to be drawn per page */
+	itemsPerPage: number
 }
 
 /** An object containing modular item configuration for an asset. Contains all of the necessary information for the
@@ -1922,7 +2415,7 @@ interface ModularItemData {
 	/** A lookup for the current page in the extended item menu for each of the item's modules */
 	pages: Record<string, number>;
 	/** A lookup for the draw data for each of the item's modules */
-	drawData: Record<string, { pageCount: number, paginate: boolean, positions: number[][] }>;
+	drawData: Record<string, ModularItemDrawData>;
 	/** A lookup for the draw functions for each of the item's modules */
 	drawFunctions: Record<string, () => void>;
 	/** A lookup for the click functions for each of the item's modules */
@@ -1944,18 +2437,28 @@ interface ModularItemData {
 		exit?: () => void,
 		validate?: ExtendedItemValidateScriptHookCallback<ModularItemOption>,
 	};
+	/**
+	 * To-be initialized properties independant of the selected item module(s).
+	 * Relevant if there are properties that are (near) exclusively managed by {@link ModularItemData.scriptHooks} functions.
+	 */
+	BaselineProperty: ItemProperties | null;
+	/**
+	 * A boolean indicating whether or not images should be drawn for the module selection screen.
+	 * Automatically generated based on {@link ModularItemModule.DrawImages} if not explicitly specified.
+	 */
+	drawImages: boolean;
 }
 
-/** A 3-tuple (or 2-tuple) containing data for drawing a button in a modular item screen. A button definition takes the
+/** A 3-tuple containing data for drawing a button in a modular item screen. A button definition takes the
  * format:
  * ```
- * [imageUrl, textKey, background]
+ * [moduleOrOption, currentOption, prefix]
  * ```
- * The imageUrl is the URL for the image that should be drawn in the button.
- * The textKey is the CSV key for the text that should be displayed in the button.
- * The background is an optional CSS color string defining the background color for the button.
+ * The moduleOrOption is the to be drawn item module or option.
+ * The currentOption is currently active option within the relevant module.
+ * The prefix is the dialog prefix for the buttons text.
  */
-type ModularItemButtonDefinition = [string, string] | [string, string, string];
+type ModularItemButtonDefinition = [ModularItemOption | ModularItemModule, ModularItemOption, string]
 
 //#endregion
 
@@ -1989,11 +2492,6 @@ interface TypedItemConfig {
 	 */
 	ChangeWhenLocked?: boolean;
 	/**
-	 * An optional validation callback function which can be used by
-	 * items to run additional validation for cases that aren't covered by configuration
-	 */
-	Validate?: ExtendedItemValidateCallback<ExtendedItemOption>;
-	/**
 	 * Contains custom dictionary entries in the event that the base ones do not suffice.
 	 */
 	Dictionary?: TypedItemDictionaryCallback[];
@@ -2010,6 +2508,11 @@ interface TypedItemConfig {
 		Validate?: ExtendedItemValidateScriptHookCallback<ExtendedItemOption>,
 		PublishAction?: ExtendedItemPublishActionCallback<ExtendedItemOption>,
 	};
+	/**
+	 * To-be initialized properties independant of the selected item module(s).
+	 * Relevant if there are properties that are (near) exclusively managed by {@link TypedItemConfig.ScriptHooks} functions.
+	 */
+	BaselineProperty?: ItemProperties;
 }
 
 interface TypedItemDialogConfig {
@@ -2036,7 +2539,7 @@ interface TypedItemDialogConfig {
 	 * A prefix for text keys for NPC dialog. This will be suffixed with the option name
 	 * to get the final NPC dialogue key (i.e. `"<npcPrefix><optionName>"`. Defaults to `"<groupName><assetName>"`
 	 */
-	NpcPrefix?: string;
+	NpcPrefix?: string | ExtendedItemNPCCallback<ExtendedItemOption>;
 }
 
 /**
@@ -2061,7 +2564,7 @@ interface TypedItemData {
 		/** The prefix used for dialog keys representing the item's chatroom messages when its type is changed */
 		chatPrefix: string | ExtendedItemChatCallback<ExtendedItemOption>;
 		/** The prefix used for dialog keys representing an NPC's reactions to item type changes */
-		npcPrefix: string;
+		npcPrefix: string | ExtendedItemNPCCallback<ExtendedItemOption>;
 	};
 	/**
 	 * An array of the chat message tags that should be included in the item's
@@ -2085,11 +2588,6 @@ interface TypedItemData {
 	 */
 	changeWhenLocked?: boolean;
 	/**
-	 * An optional validation callback function which can be used by
-	 * items to run additional validation for cases that aren't covered by configuration
-	 */
-	validate?: ExtendedItemValidateCallback<ExtendedItemOption>;
-	/**
 	 * A recond containing functions that are run on load, click, draw, exit, validate and publishaction,
 	 * with the original archetype function and parameters passed on to them. If undefined, these are ignored.
 	 * Note that scripthook functions must be loaded before `Female3DCGExtended.js` in `index.html`.
@@ -2102,6 +2600,11 @@ interface TypedItemData {
 		validate?: ExtendedItemValidateScriptHookCallback<ExtendedItemOption>,
 		publishAction?: ExtendedItemPublishActionCallback<ExtendedItemOption>,
 	};
+	/**
+	 * To-be initialized properties independant of the selected item module(s).
+	 * Relevant if there are properties that are (near) exclusively managed by {@link TypedItemData.scriptHooks} functions.
+	 */
+	BaselineProperty: ItemProperties | null;
 }
 
 /**
@@ -2141,6 +2644,8 @@ interface AppearanceUpdateParameters {
 	 * lover-only items)
 	 */
 	fromLover: boolean;
+	/** The script permission levels that the source player has with respect to the receiver */
+	permissions: ScriptPermissionLevel[];
 	/** The member number of the source player */
 	sourceMemberNumber: number;
 }
@@ -2196,6 +2701,17 @@ type VibratingItemAssetConfig = ExtendedItemAssetConfig<"vibrating", VibratingIt
 interface VibratingItemConfig {
 	/** The list of vibrator mode sets that are available on this item */
 	Options?: VibratorModeSet[];
+	/**
+	 * A record containing functions that are run on load, click, draw, exit, and validate, with the original archetype function
+	 * and parameters passed on to them. If undefined, these are ignored.
+	 * Note that scripthook functions must be loaded before `Female3DCGExtended.js` in `index.html`.
+	 */
+	ScriptHooks?: {
+		Load?: (next: () => void) => void;
+		Click?: (next: () => void) => void;
+		Draw?: (next: () => void) => void;
+		Exit?: () => void;
+	};
 }
 
 interface VibratingItemData {
@@ -2209,6 +2725,17 @@ interface VibratingItemData {
 	functionPrefix: string;
 	/** The common prefix used for all dynamic asset hook functions for the asset */
 	dynamicAssetsFunctionPrefix: string;
+	/**
+	 * A record containing functions that are run on load, click, draw, exit, and validate, with the original archetype function
+	 * and parameters passed on to them. If undefined, these are ignored.
+	 * Note that scripthook functions must be loaded before `Female3DCGExtended.js` in `index.html`.
+	 */
+	scriptHooks: {
+		load?: (next: () => void) => void;
+		click?: (next: () => void) => void;
+		draw?: (next: () => void) => void;
+		exit?: () => void;
+	};
 }
 
 /**
@@ -2272,7 +2799,7 @@ interface VariableHeightDialogConfig {
 	 * A prefix for text keys for NPC dialog. This will be suffixed with the option name
 	 * to get the final NPC dialogue key (i.e. `"<npcPrefix><optionName>"`. Defaults to `"<groupName><assetName>"`
 	 */
-	NpcPrefix?: string;
+	NpcPrefix?: string | ExtendedItemNPCCallback<ExtendedItemOption>;
 }
 
 /**
@@ -2299,7 +2826,7 @@ interface VariableHeightData {
 		/** The prefix used for dialog keys representing the item's chatroom messages when its type is changed */
 		chatPrefix: string | ExtendedItemChatCallback<ExtendedItemOption>;
 		/** The prefix used for dialog keys representing an NPC's reactions to item type changes */
-		npcPrefix: string;
+		npcPrefix: string | ExtendedItemNPCCallback<ExtendedItemOption>;
 	};
 	/**
 	 * An array of the chat message tags that should be included in the item's
@@ -2629,6 +3156,8 @@ interface PandoraBaseRoom {
 
 //#region Crafting items
 
+type CraftingMode = "Slot" | "Item" | "Property" | "Lock" | "Name" | "Color";
+
 /**
  * A struct with an items crafting-related information.
  * @see {@link Item.Craft}
@@ -2654,10 +3183,10 @@ interface CraftingItem {
 	Private: boolean;
 	/**
 	 * The type of the crafted item; only relevant for extended items and should be an empty string otherwise.
-	 * @see {@link ItemPropertiesBase.Type}
+	 * @see {@link ItemProperties.Type}
 	 */
 	Type: string | null;
-	/** An integer representing the item layering priority; see {@link ItemPropertiesBase.OverridePriority} */
+	/** An integer representing the item layering priority; see {@link ItemProperties.OverridePriority} */
 	OverridePriority: number | null;
 }
 
@@ -2683,10 +3212,10 @@ interface CraftingItemSelected {
 	/**
 	 * The type of the crafted item; only relevant for extended items and should be an empty string otherwise.
 	 * Note that `null` values, which are legal for Typed extended items, *must* be converted to empty strings.
-	 * @see {@link ItemPropertiesBase.Type}
+	 * @see {@link ItemProperties.Type}
 	 */
 	Type: string;
-	/** An integer representing the item layering priority; see {@link ItemPropertiesBase.OverridePriority} */
+	/** An integer representing the item layering priority; see {@link ItemProperties.OverridePriority} */
 	OverridePriority: number | null;
 }
 
@@ -2694,7 +3223,7 @@ interface CraftingItemSelected {
  * A struct with tools for validating {@link CraftingItem} properties.
  * @property {function} Validate - The validation function
  * @property {function} GetDefault - A function that creates default values for when the validation fails
- * @property {CraftingStatusType} - The {@link CraftingStatusCode} code for when the validation fails
+ * @property {CraftingStatusType} - The {@link CraftingStatusType} code for when the validation fails
  */
 interface CratingValidationStruct {
 	Validate: (Craft: CraftingItem, Asset: Asset | null) => boolean;
@@ -2703,3 +3232,80 @@ interface CratingValidationStruct {
 }
 
 //#endregion
+
+//#region Color
+
+/** An object defining a group of layers which can be colored together */
+interface ColorGroup {
+	/** The name of the color group */
+	name: string;
+	/** The layers contained within the color group */
+	layers: AssetLayer[];
+	/** The color index for the color group - this is the lowest color index of any of the layers within the color group */
+	colorIndex: number;
+}
+
+/**
+ * A callback function that is called when the item color UI exits
+ * @param c - The character being colored
+ * @param item - The item being colored
+ * @param save - Whether the item's appearance changes should be saved
+ */
+type itemColorExitListener = (
+	c: Character,
+	item: Item,
+	save: boolean,
+) => void;
+
+interface ItemColorStateType {
+	colorGroups: ColorGroup[];
+	colors: string[];
+	simpleMode: boolean;
+	paginationButtonX: number;
+	cancelButtonX: number;
+	saveButtonX: number;
+	colorPickerButtonX: number;
+	colorDisplayButtonX: number;
+	contentY: number;
+	groupButtonWidth: number;
+	pageSize: number;
+	pageCount: number;
+	colorInputWidth: number;
+	colorInputX: number;
+	colorInputY: number;
+	exportButtonX: number;
+	importButtonX: number;
+	resetButtonX: number;
+	drawImport: () => Promise<string>;
+	drawExport: (data: string) => Promise<void>;
+}
+
+//#end region
+
+// #region property
+
+// NOTE: Use the intersection operator to enforce that the it remains a `keyof ItemProperties` subtype
+/** Property keys of {@link ItemProperties} with text input fields */
+type PropertyTextNames = keyof ItemProperties & (
+	"Text" | "Text2" | "Text3"
+);
+
+/**
+ * A callback signature for handling (throttled) text changes.
+ * @param {Character} C - The character being modified
+ * @param {Item} item - The item being modified
+ * @param {PropertyTextNames} PropName - The property wherein the updated text should be stored
+ * @param {string} Text - The new text to be assigned to the item
+ * @returns {void} Nothing
+ */
+type PropertyTextEventListener = (
+	C: Character,
+	Item: Item,
+	PropName: PropertyTextNames,
+	Text: string,
+) => void;
+
+/** A record type with custom event listeners for one or more text input fields. */
+type PropertyTextEventListenerRecord = Partial<Record<PropertyTextNames, PropertyTextEventListener>>;
+
+// #end region